% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enum_param_property.R
\name{as_param_property}
\alias{as_param_property}
\title{Convert vector to factor with standard parameter property levels}
\usage{
as_param_property(x, levels = enum_param_property())
}
\arguments{
\item{x}{character vector}

\item{levels}{optional vector of levels (default = enum_param_property())}
}
\value{
factor with levels in standard order
}
\description{
Convert vector to factor with standard parameter property levels
}
