% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enum_param_scale.R
\name{as_param_scale}
\alias{as_param_scale}
\title{Convert vector to factor with standard parameter scale levels}
\usage{
as_param_scale(x, levels = enum_param_scale())
}
\arguments{
\item{x}{character vector}

\item{levels}{optional vector of levels (default = enum_param_scale())}
}
\value{
factor with levels in standard order
}
\description{
Convert vector to factor with standard parameter scale levels
}
