% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enum_param_time_aspect.R
\name{as_param_time_aspect}
\alias{as_param_time_aspect}
\title{Convert vector to factor with standard parameter time aspect levels}
\usage{
as_param_time_aspect(x, levels = enum_param_time_aspect())
}
\arguments{
\item{x}{character vector}

\item{levels}{optional vector of levels (default = enum_param_time_aspect())}
}
\value{
factor with levels in standard order
}
\description{
Convert vector to factor with standard parameter time aspect levels
}
