% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enum_specimen_type.R
\name{as_specimen_type}
\alias{as_specimen_type}
\title{Convert vector to factor with standard specimen type levels}
\usage{
as_specimen_type(x, levels = enum_specimen_type())
}
\arguments{
\item{x}{character vector}

\item{levels}{optional vector of levels (default = enum_specimen_type())}
}
\value{
factor with levels in standard order
}
\description{
Convert vector to factor with standard specimen type levels
}
