% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bar_pos_neg.R
\name{bar_chart_pos_neg}
\alias{bar_chart_pos_neg}
\title{Add a positive to negative bar chart to a reactable table}
\usage{
bar_chart_pos_neg(
  label,
  value,
  max_value = 1,
  height = "16px",
  pos_fill = "#008097",
  neg_fill = "#ec008c"
)
}
\arguments{
\item{label}{column name containing an address string}
}
\value{
Cell theme for a reactable table
}
\description{
Add a positive to negative bar chart to a reactable table
}
\examples{
test <- "test"
}
