% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ckan_query.R
\name{ckan_query}
\alias{ckan_query}
\title{Function to pull data using SQL from CKAN}
\usage{
ckan_query(sql_query, geometry = FALSE, nested = FALSE)
}
\arguments{
\item{sql_query}{A string SQL query}

\item{geometry}{Boolean indicating if dataframe contains geometry or not. If TRUE, function returns a geometry object ready for mapping}

\item{nested}{Boolean indicating if function will return list including metadata and records. If FALSE, only records are returned}
}
\description{
\code{ckan_query} allows users to pull data using a custom SQL query, with the option of adding geometry if present in the dataset. This function returns either a dataframe or a list including metadata and records.
}
