% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpal-package.R
\docType{package}
\name{cpal-package}
\alias{cpal}
\alias{cpal-package}
\title{cpal: Standardized Reference Tables from CPAL}
\description{
Provides access to CPAL (Catálogo Português de Análises de Laboratório) ontology datasets including blood groups, antimicrobials, specimens, and other clinical pathology terminology.
}
\author{
\strong{Maintainer}: Nuno Aguileira \email{nunoaguileira1@gmail.com}

}
\keyword{internal}
