% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{cpal_ab0_group_phenotype}
\alias{cpal_ab0_group_phenotype}
\title{CPAL AB0 Group Phenotype Data}
\format{
A data frame with 4 obs. of 11 variables:
\describe{
\item{Line code | Código de linha}{An identifier for the line within the original data source.}
\item{OID Code System | Sistema de Codificação OID}{The Object Identifier (OID) code system used.}
\item{Code | Código}{The specific code for the AB0 group phenotype.}
\item{Fully Specified Name (FSN) EN | Termo Totalmente Especificado EN}{The fully specified name of the phenotype in English.}
\item{Preferred Term (PT) EN | Termo de Preferência EN}{The preferred term for the phenotype in English.}
\item{Preferred Term (PT) PT | Termo de Preferência PT}{The preferred term for the phenotype in Portuguese.}
\item{Synonym 1 | Sinónimo 1}{A first synonym for the phenotype term.}
\item{Synonym 2 | Sinónimo 2}{A second synonym for the phenotype term.}
\item{Version | Versão}{The release date of the data version.}
\item{Type of Change | Tipo Alteração}{Describes the type of change made to the record.}
\item{Type of Change Code | Código Tipo Alteração}{The code corresponding to the type of change.}
}
}
\source{
Catálogo Português de Análises de Laboratório (CPAL), available in \url{https://www.ctc.min-saude.pt/catalogos/analises-de-laboratorio/}.
}
\usage{
cpal_ab0_group_phenotype
}
\description{
This dataset contains information related to AB0 blood group phenotypes from CPAL (Catálogo Português de Análises de Laboratório), including various codes, names, and versioning details.
}
\keyword{datasets}
