% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{cpal_antimicrobial}
\alias{cpal_antimicrobial}
\title{CPAL Antimicrobial Data}
\format{
A data frame with 232 obs. of 11 variables:
\describe{
\item{Line code | Código de linha}{An identifier for the line within the original data source.}
\item{Disposition - Code | Código - Disposição}{A code indicating the disposition or status of the antimicrobial record.}
\item{OID Code System | Sistema de Codificação OID}{The Object Identifier (OID) code system used for the classification.}
\item{Code LOINC| Código LOINC}{The LOINC code associated with the antimicrobial.}
\item{Long Common Name (LCN) | Nome Longo Comum (NLC)}{The long common name of the antimicrobial.}
\item{Preferred Term (PT) EN | Termo de Preferência EN}{The preferred term for the antimicrobial in English.}
\item{Preferred Term (PT) PT | Termo de Preferência PT}{The preferred term for the antimicrobial in Portuguese.}
\item{Synonym 1 | Sinónimo 1}{A first synonym for the antimicrobial term.}
\item{Version | Versão}{The release date of the data version.}
\item{Type of Change | Tipo Alteração}{Describes the type of change made to this record.}
\item{Type of Change Code | Código Tipo Alteração}{The code corresponding to the type of change.}
}
}
\source{
Catálogo Português de Análises de Laboratório (CPAL), available in \url{https://www.ctc.min-saude.pt/catalogos/analises-de-laboratorio/}.
}
\usage{
cpal_antimicrobial
}
\description{
This dataset provides detailed information on various antimicrobials, including their codes and names, from CPAL (Catálogo Português de Análises de Laboratório).
}
\keyword{datasets}
