% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{cpal_antimicrobial_type}
\alias{cpal_antimicrobial_type}
\title{CPAL Antimicrobial Type Data}
\format{
A data frame with 5 obs. of 9 variables:
\describe{
\item{Line code | Código de linha}{An identifier for the line within the original data source.}
\item{OID Code System | Sistema de Codificação OID}{The Object Identifier (OID) code system used.}
\item{Disposition - Code SNOMED CT | Código SNOMED CT - Disposição}{The SNOMED CT code related to the disposition of the antimicrobial type.}
\item{Fully Specified Name (FSN) EN | Termo Totalmente Especificado EN}{The fully specified name of the antimicrobial type in English.}
\item{Preferred Term (PT) EN | Termo de Preferência EN}{The preferred term for the antimicrobial type in English.}
\item{Preferred Term (PT) PT | Termo de Preferência PT}{The preferred term for the antimicrobial type in Portuguese.}
\item{Version | Versão}{The release date of the data version.}
\item{Type of Change | Tipo Alteração}{Describes the type of change made to the record.}
\item{Type of Change Code | Código Tipo Alteração}{The code corresponding to the type of change.}
}
}
\source{
Catálogo Português de Análises de Laboratório (CPAL), available in \url{https://www.ctc.min-saude.pt/catalogos/analises-de-laboratorio/}.
}
\usage{
cpal_antimicrobial_type
}
\description{
This dataset contains classifications and details about different types of antimicrobials from CPAL (Catálogo Português de Análises de Laboratório).
}
\keyword{datasets}
