% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{cpal_clinical_pathology_procedure}
\alias{cpal_clinical_pathology_procedure}
\title{CPAL Clinical Pathology Procedure Data}
\format{
A data frame with 6857 obs. of 20 variables:
\describe{
\item{Line code | Código de linha}{An identifier for the line within the original data source.}
\item{Área}{The main area or category of the clinical pathology procedure.}
\item{Sub Área}{A sub-category or sub-area of the clinical pathology procedure.}
\item{Code SNS| Código SNS}{The SNS (Serviço Nacional de Saúde) code associated with the procedure.}
\item{Code Conv. | Código Conv.}{The conventional code associated with the procedure.}
\item{Long Common Name OM/SNS/Convencionados | Termo de Preferência OM/SNS/Convencionados}{The long common name or preferred term for the procedure, as used by OM/SNS/Conventioned entities.}
\item{Order Code | Código Pedido}{The code used for ordering the procedure.}
\item{Preferred Term (PT) PT | Termo de Preferência PT...8}{The preferred term for the procedure in Portuguese. Note: This column name includes an automatic suffix due to a duplicate header in the source data.}
\item{Preferred Term (PT) EN | Termo de Preferência EN}{The preferred term for the procedure in English.}
\item{Code LOINC | Código LOINC Resultado}{The LOINC code associated with the result of the procedure.}
\item{Preferred Term (PT) PT | Termo de Preferência PT...11}{The preferred term for the procedure in Portuguese. Note: This column name includes an automatic suffix due to a duplicate header in the source data.}
\item{Long Common Name (LCN) | Nome Longo Comum (NLC)}{The long common name of the procedure.}
\item{Synonym | Sinónimo}{A synonym for the procedure term.}
\item{Property | Propriedade}{The property or characteristic measured by the procedure.}
\item{Scale | Escala}{The scale of measurement.}
\item{Method | Método}{The method used for the clinical pathology procedure.}
\item{UCUM Units Example | Exemlo Unidades UCUM}{An example of UCUM (Unified Code for Units of Measure) units relevant to the procedure.}
\item{Version| Versão}{The release date of the data version.}
\item{Type of Change | Tipo Alteração}{Describes the type of change made to this record.}
\item{Type of Change Code | Código Tipo Alteração}{The code corresponding to the type of change.}
}
}
\source{
Catálogo Português de Análises de Laboratório (CPAL), available in \url{https://www.ctc.min-saude.pt/catalogos/analises-de-laboratorio/}.
}
\usage{
cpal_clinical_pathology_procedure
}
\description{
This dataset contains detailed information on various clinical pathology procedures, including their codes, names, properties, and change history, from CPAL (Catálogo Português de Análises de Laboratório).
}
\keyword{datasets}
