% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{cpal_color_of_specimen}
\alias{cpal_color_of_specimen}
\title{CPAL Color of Specimen Data}
\format{
A data frame with 17 obs. of 10 variables:
\describe{
\item{Line code | Código de linha}{An identifier for the line within the original data source.}
\item{OID Code System | Sistema de Codificação OID}{The Object Identifier (OID) code system used for the classification.}
\item{Code SNOMED CT | Código SNOMED CT}{The SNOMED CT code representing the specific color of the specimen.}
\item{Fully Specified Name (FSN) EN | Termo Totalmente Especificado EN}{The fully specified name of the specimen color in English.}
\item{Preferred Term (PT) EN | Termo de Preferência EN}{The preferred term for the specimen color in English.}
\item{Preferred Term (PT) PT | Termo de Preferência PT}{The preferred term for the specimen color in Portuguese.}
\item{Synonym 1 | Sinónimo 1}{A first synonym for the specimen color term.}
\item{Version | Versão}{The release date of the data version.}
\item{Type of Change | Tipo Alteração}{Describes the type of change made to this record.}
\item{Type of Change Code | Código Tipo Alteração}{The code corresponding to the type of change.}
}
}
\source{
Catálogo Português de Análises de Laboratório (CPAL), available in \url{https://www.ctc.min-saude.pt/catalogos/analises-de-laboratorio/}.
}
\usage{
cpal_color_of_specimen
}
\description{
This dataset contains classifications and details about the colors of biological specimens from CPAL (Catálogo Português de Análises de Laboratório).
}
\keyword{datasets}
