% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{cpal_crystal}
\alias{cpal_crystal}
\title{CPAL Crystal Data}
\format{
A data frame with 32 obs. of 9 variables:
\describe{
\item{Line code | Código de linha}{An identifier for the line within the original data source.}
\item{OID Code System | Sistema de Codificação OID}{The Object Identifier (OID) code system used for the classification.}
\item{Code SNOMED CT | Código SNOMED CT}{The SNOMED CT code representing the specific type of crystal.}
\item{Fully Specified Name (FSN) EN | Termo Totalmente Especificado EN}{The fully specified name of the crystal type in English.}
\item{Preferred Term (PT) EN | Termo de Preferência EN}{The preferred term for the crystal type in English.}
\item{Synonym 1 | Sinónimo 1}{A first synonym for the crystal type term.}
\item{Version | Versão}{The release date of the data version.}
\item{Type of Change | Tipo Alteração}{Describes the type of change made to this record.}
\item{Type of Change Code | Código Tipo Alteração}{The code corresponding to the type of change.}
}
}
\source{
Catálogo Português de Análises de Laboratório (CPAL), available in \url{https://www.ctc.min-saude.pt/catalogos/analises-de-laboratorio/}.
}
\usage{
cpal_crystal
}
\description{
This dataset contains classifications and details about various types of crystals found in clinical pathology specimens from CPAL (Catálogo Português de Análises de Laboratório).
}
\keyword{datasets}
