% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{cpal_laboratory_study_type}
\alias{cpal_laboratory_study_type}
\title{CPAL Laboratory Study Type Data}
\format{
A data frame with 20 obs. of 11 variables:
\describe{
\item{Line code | Código de linha}{An identifier for the line within the original data source.}
\item{OID Code System | Sistema de Codificação OID}{The Object Identifier (OID) code system used for the classification.}
\item{Code LOINC | Código LOINC}{The LOINC code associated with the laboratory study type.}
\item{Long Common Name (LCN) | Nome Longo Comum (NLC)}{The long common name of the laboratory study type.}
\item{Preferred Term (PT) PT | Termo de Preferência PT}{The preferred term for the laboratory study type in Portuguese.}
\item{Synonym 1 | Sinónimo 1}{A first synonym for the laboratory study type term.}
\item{Synonym 2 | Sinónimo 2}{A second synonym for the laboratory study type term.}
\item{Synonym 3 | Sinónimo 3}{A third synonym for the laboratory study type term.}
\item{Version | Versão}{The release date of the data version.}
\item{Type of Change | Tipo Alteração}{Describes the type of change made to this record.}
\item{Type of Change Code | Código Tipo Alteração}{The code corresponding to the type of change.}
}
}
\source{
Catálogo Português de Análises de Laboratório (CPAL), available in \url{https://www.ctc.min-saude.pt/catalogos/analises-de-laboratorio/}.
}
\usage{
cpal_laboratory_study_type
}
\description{
This dataset contains classifications and details about various types of laboratory studies from CPAL (Catálogo Português de Análises de Laboratório).
}
\keyword{datasets}
