% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpal_palette.R
\name{cpal_palette}
\alias{cpal_palette}
\title{Generates a color palette based on CPAL standards}
\usage{
cpal_palette(
  n,
  name = c("cpal_ten", "cpal_fiv", "cpal_tri", "cpal_div", "cpal_two")
)
}
\arguments{
\item{n}{total number of hex values to be generated by palette}

\item{name}{name of color palette to generate hex values from}
}
\value{
A dataframe with a list n hex values for the chosen color palette
}
\description{
Generates a color palette based on CPAL standards
}
\examples{
div_OrTe <- cpal_palette(n = 30, name = "cpal_div")
par(mar = rep(0, 4))
pie(rep(1, length(div_OrTe)), col = div_OrTe)
}
