% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpal_release.R
\name{cpal_release}
\alias{cpal_release}
\title{CPAL Release Date}
\usage{
cpal_release()
}
\value{
A \link[base:Dates]{Date} object.
}
\description{
This function retrieves the release date of the bundled
CPAL terminology from the package version.
}
\examples{
cpal_release()

}
