% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{cpal_rhesus_factor}
\alias{cpal_rhesus_factor}
\title{CPAL Rhesus Factor Data}
\format{
A data frame with 2 obs. 11 variables:
\describe{
\item{Line code | Código de linha}{An identifier for the line within the original data source.}
\item{OID Code System | Sistema de Codificação OID}{The Object Identifier (OID) code system used for the classification.}
\item{Code SNOMED CT | Código SNOMED CT}{The SNOMED CT code representing the specific Rhesus factor.}
\item{Fully Specified Name (FSN) EN | Termo Totalmente Especificado EN}{The fully specified name of the Rhesus factor in English.}
\item{Preferred Term (PT) EN | Termo de Preferência EN}{The preferred term for the Rhesus factor in English.}
\item{Preferred Term (PT) PT | Termo de Preferência PT}{The preferred term for the Rhesus factor in Portuguese.}
\item{Synonym 1 | Sinónimo 1}{A first synonym for the Rhesus factor term.}
\item{Synonym 2 | Sinónimo 2}{A second synonym for the Rhesus factor term.}
\item{Version | Versão}{The release date of the data version.}
\item{Type of Change | Tipo Alteração}{Describes the type of change made to this record.}
\item{Type of Change Code | Código Tipo Alteração}{The code corresponding to the type of change.}
}
}
\source{
Catálogo Português de Análises de Laboratório (CPAL), available in \url{https://www.ctc.min-saude.pt/catalogos/analises-de-laboratorio/}.
}
\usage{
cpal_rhesus_factor
}
\description{
This dataset contains classifications and details about Rhesus (Rh) blood factors from CPAL (Catálogo Português de Análises de Laboratório).
}
\keyword{datasets}
