% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{cpal_version}
\alias{cpal_version}
\title{CPAL Version Data}
\format{
A data frame with 17 obs. of 5 variables:
\describe{
\item{Versão}{The version number or identifier of the CPAL data release.}
\item{Data}{The release date of the specific CPAL data version.}
\item{Estado}{The status of the data version.}
\item{Responsável}{The entity or individual responsible for the version release.}
\item{Alterações}{A description of the changes and updates included in this data version.}
}
}
\source{
Catálogo Português de Análises de Laboratório (CPAL), available in \url{https://www.ctc.min-saude.pt/catalogos/analises-de-laboratorio/}.
}
\usage{
cpal_version
}
\description{
This dataset provides a historical log of the different versions and updates of the CPAL (Catálogo Português de Análises de Laboratório) data.
}
\keyword{datasets}
