% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enum_param_scale.R
\name{enum_param_scale}
\alias{enum_param_scale}
\title{Enumerate parameter scales}
\usage{
enum_param_scale()
}
\value{
character vector of parameter scales in standard order
}
\description{
Enumerate parameter scales
}
