% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enum_param_time_aspect.R
\name{enum_param_time_aspect}
\alias{enum_param_time_aspect}
\title{Enumerate parameter time aspects}
\usage{
enum_param_time_aspect()
}
\value{
character vector of time aspects in standard order
}
\description{
Enumerate parameter time aspects
}
