% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enum_specimen_type.R
\name{enum_specimen_type}
\alias{enum_specimen_type}
\title{Enumerate specimen types}
\usage{
enum_specimen_type()
}
\value{
character vector of specimen types in standard order
}
\description{
Enumerate specimen types
}
