% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_city.R
\name{extract_city}
\alias{extract_city}
\title{Extract city name from an address string}
\usage{
extract_city(field = NULL)
}
\arguments{
\item{field}{column name containing an address string}
}
\value{
A column that should only contain the city name from an address string
}
\description{
Extract city name from an address string
}
