% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_zip.R
\name{extract_zip}
\alias{extract_zip}
\title{Extract zip code from an address string}
\usage{
extract_zip(field = NULL)
}
\arguments{
\item{field}{column name containing an address string}
}
\value{
A column that should only contain the zip code from an address string
}
\description{
Extract zip code from an address string
}
