% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firearms.R
\docType{data}
\name{firearms}
\alias{firearms}
\title{List of firearms used in Dallas Police Department Police Incidents data.}
\format{
A data frame with 8 rows and 1 variables:
\describe{
\item{firearm}{a gun type}
...
}
}
\usage{
firearms
}
\description{
A dataset containing eight firearm types
}
\keyword{datasets}
