% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/font_cpal.R
\name{font_cpal}
\alias{font_cpal}
\title{CPAL font install}
\usage{
font_cpal(x)
}
\arguments{
\item{x}{placeholder}
}
\value{
Installs all fonts currently used by the Child Poverty Action Lab
}
\description{
CPAL font install
}
\examples{
ggplot2::ggplot(iris, ggplot2::aes(x=Petal.Length, y=Sepal.Width, color = Species)) +
ggplot2::geom_point() +
theme_cpal()

}
