% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeoIntersect.R
\name{GeoIntersect}
\alias{GeoIntersect}
\title{Intersection of two sf geographies used to summarize variables between each}
\usage{
GeoIntersect(polylg = NULL, polysm = NULL)
}
\arguments{
\item{polylg}{The larger sf geography that can be summarized down into a smaller geography}

\item{polysm}{The smaller sf geography that can be summarized up into a larger geography}
}
\value{
An sf dataframe that has been intersected between both geographies and now contains the calculated area of all geographies
and the percentage of intersection for the small and large geographies.
}
\description{
Intersection of two sf geographies used to summarize variables between each
}
