% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_acs_years.R
\name{get_acs_years}
\alias{get_acs_years}
\title{Pull ACS Census data for multiple years across the same geographies and combine into a single dataframe.}
\usage{
get_acs_years(
  variables = NULL,
  years = NULL,
  geography = NULL,
  state = NULL,
  survey = NULL
)
}
\arguments{
\item{variables}{Character string or vector of character strings of variable IDs. tidycensus automatically returns the estimate and the margin of error associated with the variable.}

\item{years}{a dataframe containing a list of the years to include data from.}

\item{geography}{The geography of your data.}

\item{state}{An optional vector of states for which you are requesting data. State names, postal codes, and FIPS codes are accepted. Defaults to NULL.}

\item{survey}{The ACS contains one-year, three-year, and five-year surveys expressed as "acs1", "acs3", and "acs5". The default selection is "acs5."}
}
\value{
A dataframe containing all identified census variables for the specified years.
}
\description{
Pull ACS Census data for multiple years across the same geographies and combine into a single dataframe.
}
