% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ntx_county.R
\docType{data}
\name{ntx_county}
\alias{ntx_county}
\title{List of North Texas counties surrounding Dallas County.}
\format{
A data frame with 8 rows and 1 variables:
\describe{
\item{county}{a gun type}
...
}
}
\usage{
ntx_county
}
\description{
A dataset containing eight county names
}
\keyword{datasets}
