% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_cpal.R
\name{theme_cpal}
\alias{theme_cpal}
\title{CPAL ggplot theme}
\usage{
theme_cpal(base_size = 36, base_family = "sans")
}
\arguments{
\item{base_size}{text size for chart theme}

\item{base_family}{font family for chart theme}
}
\value{
A ggplot base theme based on CPAL standards
}
\description{
CPAL ggplot theme
}
\examples{
ggplot2::ggplot(iris, ggplot2::aes(x=Petal.Length, y=Sepal.Width, color = Species)) +
ggplot2::geom_point() +
theme_cpal()
}
