% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeoIntersect.R
\name{GeoIntersect}
\alias{GeoIntersect}
\title{Intersection of two sf geographies used to summarize variables between each}
\usage{
GeoIntersect(x, y)
}
\arguments{
\item{x}{The larger sf geography that can be summarized down into a smaller geography}

\item{y}{The smaller sf geography that can be summarized up into a larger geography}
}
\value{
An sf dataframe that has been intersected between both geographies and now contains the calculated area of all geographies
and the percentage of intersection for the small and large geographies.
}
\description{
Intersection of two sf geographies used to summarize variables between each
}
\examples{
TexasCD <- tigris::congressional_districts(state = "TX")
plot(TexasCD["geometry"])

DallasCounty <- tigris::county_subdivisions(state = "TX", county = "Dallas County")
plot(DallasCounty["geometry"])

DallasCD <- TexasCD[DallasCounty, ]

DallasInt <- GeoIntersect(x = DallasCounty, y = DallasCD)

plot(DallasInt["geometry"])
}
